first step of the refactor renaming the main file to something that makes sense
